app/requirements.txt
Flask==2.0.1
virtualenv==20.4.7
psycopg2-binary==2.8.6
